+++ App/App/Assets.xcassets/AccentColor.colorset/Contents.json
