+++ App/App/Assets.xcassets/AppIcon.appiconset/Contents.json
