+++ App/App/MainApp.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+
+@main
+struct MainApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
