+++ OpenLogin.xcworkspace/xcshareddata/swiftpm/Package.resolved
