+++ OpenLogin/.gitignore
@@ -1,0 +1,9 @@
+# Swift Package Manager
+
+.DS_Store
+/.build
+/Packages
+/*.xcodeproj
+xcuserdata/
+DerivedData/
+.swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata
