+++ OpenLogin/Sources/OpenLogin/Module.swift
@@ -1,0 +1,1 @@
+import Foundation
