--- Sources/OpenloginSwiftSdk/Types.swift
@@ -1,12 +1,0 @@
-import Foundation
-
-public enum Network : String {
-    case mainnet = "mainnet"
-    case testnet = "testnet"
-    case development = "development"
-}
-
-public enum Method : String {
-    case login = "/login"
-    case logout = "/logout"
-}
