--- Tests/LinuxMain.swift
@@ -1,7 +1,0 @@
-import XCTest
-
-import OpenloginSwiftSdkTests
-
-var tests = [XCTestCaseEntry]()
-tests += OpenloginSwiftSdkTests.allTests()
-XCTMain(tests)
