--- Tests/OpenloginSwiftSdkTests/XCTestManifests.swift
@@ -1,9 +1,0 @@
-import XCTest
-
-#if !canImport(ObjectiveC)
-public func allTests() -> [XCTestCaseEntry] {
-    return [
-        testCase(OpenloginSwiftSdkTests.allTests),
-    ]
-}
-#endif
